### 2024-01-23 #6272 @editorialbot
[REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker
**Submitting author:** @SarahMaman (<a href="http://orcid.org/0000-0001-6453-902X">MAMAN Sarah</a>)
**Repository:** https://github.com/Bios4Biol/intronSeeker
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.0
