### 2024-02-13 #6272 @CFGrote
The code's repo (https://forgemia.inra.fr/bios4biol/intronseeker) is not on a public repo hosting service, so I cannot open issues nor push pull requests. This severely limits options for the community to contribute to or comment on the code. I would strongly suggest to move the repo to a public service such as github.com, bitbucket.org, gitlab.com. @SarahMaman, can you comment?
