### 2024-02-15 #6272 @SarahMaman
Hello,
Thank you for your feedback.
We are going to add a site on github for issues so that people can post their issues without being connected to the forge-mia site, they just need a github account. And, if possible, we'll leave the code on forge-mia .

> The code's repo (https://forgemia.inra.fr/bios4biol/intronseeker) is not on a public repo hosting service, so I cannot open issues nor push pull requests. This severely limits options for the community to contribute to or comment on the code. I would strongly suggest to move the repo to a public service such as github.com, bitbucket.org, gitlab.com. @SarahMaman, can you comment?
