### 2024-02-15 #6272 @tkchafin
One of the submission requirements for JOSS is "The software must be hosted at a location where users can open issues and propose code changes without manual approval of (or payment for) accounts."

I think the proposed solution to separate the code from the public access point would be confusing, and also still not provide a mechanism for users to propose changes (i.e., via forking and a pull request). Contributing guidelines are something else that will come up during the review -- is there a reason why the code needs to be hosted on the private gitlab?
