### 2024-02-15 #6272 @SarahMaman
> @SarahMaman indeed our requirements state "`The software must be hosted at a location where users can open issues and propose code changes without manual approval of (or payment for) accounts.`". Please ensure your software is hosted on a public repository and that forking/posting issues are enabled. If you are unable to comply with this requirement we will unfortunately need to reject this submission.

We understand, we'll export intronSeeker to GitHub.
