### 2024-02-26 91a2d52
Add anaconda clien install command line
