### 2024-02-26 #6272 @SarahMaman
> # Functionality/Installation
>
> I followed the instructions in Install.md and hit this roadblock (after cloning the repo)
>
> ```
> ▶ conda install -n base --override-channels -c conda-forge mamba 'python_abi=*=*cp*'
> ...
> -> ok
>
> ▶ cd intronseeker
>
> ▶ CONDA_SOLVER="libmamba" /bin/bash ./setup.sh
>
> ### Begin installing intronSeeker...
>
> ### Conda environment installing...
>
>
> SpecNotFound: Anaconda Client is required to interact with anaconda.org or an Anaconda API. Please run `conda install anaconda-client -n base`.
>
> FAIL : Conda environment installation FAILED.
> ```

Hello, I have added this command `conda install anaconda-client -n base` to the installation documentation. Thanks a lot.
