+++ data/REAL_DATA/Azolla_f_SRR1618559/README.md
@@ -58,4 +58,4 @@ intronSeeker buildReport  -F --config_file  SRR1618559.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Azolla_f_GBTV01.html
+HTML report is available in public directory and here : http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Azolla_f_GBTV01.html
