### 2024-03-12 Bios4Biol/intronSeeker#1 @tkchafin
[JOSS REVIEW] Installation fails
This is a part of the JOSS review (https://github.com/openjournals/joss-reviews/issues/6272)

---

On first attempt I get an issue with missing anaconda-client

```
(base) tyler@Tylers-MacBook-Pro-2 intronSeeker % CONDA_SOLVER="libmamba" /bin/bash ./setup.sh

### Begin installing intronSeeker...

### Conda environment installing...

SpecNotFound: Anaconda Client is required to interact with anaconda.org or an Anaconda API. Please run `conda install anaconda-client -n base`.

FAIL : Conda environment installation FAILED.
```

If I manually install anaconda-client I get an issue about needing to log in, but I think this is because `setup.sh` is not correctly resolving the path to the `environment.yml` file:
```
### Begin installing intronSeeker...

### Conda environment installing...

SpecNotFound: (3)/Downloads/intronSeeker/config/environment.yml was not found on anaconda.org.
You may need to be logged in. Try running:
    anaconda login

FAIL : Conda environment installation FAILED.
```

Output of `pwd`:
```
(base) tyler@Tylers-MacBook-Pro-2 intronSeeker % pwd
/Users/tyler/Downloads/intronSeeker
```

So I'm not sure where the "(3)/" is coming from. Is it not possible to make this installable directly via conda? Like:
`conda -c YOUR_CHANNEL intronSeeker`
