+++ paper/paper.md
@@ -67,26 +67,26 @@ The twenty public TSA contigs sets processed by IntronSeeker are classified in t
 
 | Species     |  (super)kingdom  | TSA id |Nb contigs | SRA and link to HTML intronSeeker report                                        |     Nb reads. |   \% cwi|
 | :---       |  :----:   |   :----: |:----:  |    :----:  |      :--- |    :--- |
-| Salvia m.    |  Plantae   |  GJJN01 | 69 705 |  [SRR15718805](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Salvia_m_GJJN01.html) |   23 086 599 | **11.96%**|
-| Platichthys s.  |  Animalia  |  GAPK01 | 30 630 |  [SRR1023744](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Platichthys_s_GAPK01.html) | 516 791 904  | **10,71%**|
-| Rigidoporus m.  |  Fungi    |  GDMN01 | 34 441 |  [SRR2187438](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Rigidoporus_m_GDMN01.html)  |   75 600 628 |  **5,97%**|
-| Isatis t.    |  Plantae   |  GARR01 | 33 238 | [SRR1051997](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Isatis_t_GARR01.html)  |  113 134 348 |  **5,41%**|
-| Goniomonas a.  |  Eukaryota  |  GGUN01 | 48 844 | [SRR7601946](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Goniomonas_a_GGUN01.html)   |  82 416 944 |  **5,27%**|
-| Vriesea c.    |  Plantae   |  GHCB01 | 41 228 | [SRR500874](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Vriesea_c_GHCB01.html)   |  85 726 288  | **3,94%**|
-| Graminella n.  |  Animalia  |  GAQX01 | 37 537 | [SRR857257](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Graminella_n_GAQX01.html)   |  43 693 708  | **2,88%**|
-| Carassius g.   |  Animalia  |  GJKR01 | 109 966 | [SRR12596368](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Carassius_g_GJKR01.html)  |  21 661 960 | **2.72%**|
-| Rhizopus a.   |  Fungi    |  GDUK01 | 30 601 | [SRR2104505](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Rhizopus_a_GDUK01.html)   |  64 801 576  | **2,51%**|
-| Diplonema p.   | Eukaryota  | GJNJ01  | 114 037 | [SRR14933372](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Diplonema_p_GJNJ01.html)  | 63 775 926   |  **2.35%**|
-| Azolla f.    | Plantae   | GBTV01  | 36 091 | [SRR1618559]( https://bios4biol.pages.mia.inra.fr/intronseeker/report_Azolla_f_GBTV01.html)  |  122 059 452 |  **1,97%**|
-| Tripidium r.   | Plantae   | GJDA01  | 106 494 | [SRR14143372]( https://bios4biol.pages.mia.inra.fr/intronseeker/report_Tripidium_r_GJDA01.html)  | 15 357 748  | **1.50%**|
-| Cimex l.     |  Animalia  | GBYH01  | 39 124 | [SRR1660436](https://bios4biol.pages.mia.inra.fr/intronseeker/report_Cimex_l_GBYH01.html)   | 329 875 624  |  **1,24%**|
[…]
