+++ data/GBS-GTF_Based_Simulation/Caenorhabditis_elegans/README.md
@@ -85,4 +85,4 @@ force: -F
 intronSeeker buildReport  -F --config_file  Cele.cfg;
 ```
 
-HTML report is available in intronSeeker public directory : https://bios4biol.pages.mia.inra.fr/intronseeker/report_Cele.html
+HTML report is available in intronSeeker public directory : http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Cele.html
