+++ data/REAL_DATA/Carassius_g_SRR12596368/README.md
@@ -60,4 +60,4 @@ intronSeeker buildReport -F --config_file  SRR12596368.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Carassius_g_GJKR01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Carassius_g_GJKR01.html
