+++ data/REAL_DATA/Choromytilus_c_SRR15058678/README.md
@@ -59,4 +59,4 @@ intronSeeker buildReport -F --config_file  SRR15058678.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Choromytilus_c_GJJD01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Choromytilus_c_GJJD01.html
