+++ data/REAL_DATA/Cimex_l_SRR1660436/README.md
@@ -57,4 +57,4 @@ intronSeeker buildReport -F --config_file  SRR1660436.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Cimex_l_GBYH01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Cimex_l_GBYH01.html
