+++ data/REAL_DATA/Diplonema_p_SRR14933372/README.md
@@ -60,4 +60,4 @@ intronSeeker buildReport -F --config_file  SRR14933372.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Diplonema_p_GJNJ01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Diplonema_p_GJNJ01.html
