+++ data/REAL_DATA/Graminella_n_SRR857257/README.md
@@ -60,4 +60,4 @@ intronSeeker buildReport -F --config_file  SRR857257.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Graminella_n_GAQX01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Graminella_n_GAQX01.html
