+++ data/REAL_DATA/Isatis_t_SRR1051997/README.md
@@ -59,4 +59,4 @@ intronSeeker buildReport -F --config_file  SRR1051997.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Isatis_t_GARR01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Isatis_t_GARR01.html
