+++ data/REAL_DATA/Phytophthora_c_SRR1206033/README.md
@@ -57,4 +57,4 @@ intronSeeker buildReport -F --config_file  SRR1206033.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Phytophtora_c_GBGX01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Phytophtora_c_GBGX01.html
