+++ data/REAL_DATA/Piromyces_sp_SRR7819335/README.md
@@ -59,5 +59,5 @@ intronSeeker buildReport -F --config_file  SRR7819335.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Piromyces_sp_GGXH01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Piromyces_sp_GGXH01.html
 
