+++ data/REAL_DATA/Platichthys_s_SRR1023744/README.md
@@ -59,5 +59,5 @@ intronSeeker buildReport -F --config_file  SRR1023744.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Platichthys_s_GAPK01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Platichthys_s_GAPK01.html
 
