+++ data/REAL_DATA/Rhizopus_a_SRR2104505/README.md
@@ -59,4 +59,4 @@ intronSeeker buildReport -F --config_file  SRR2104505.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Rhizopus_a_GDUK01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Rhizopus_a_GDUK01.html
