+++ data/REAL_DATA/Rigidoporus_m_SRR2187438/README.md
@@ -60,5 +60,5 @@ intronSeeker buildReport -F --config_file  SRR2187438.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Rigidoporus_m_GDMN01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Rigidoporus_m_GDMN01.html
 
