+++ data/REAL_DATA/Tripidium_r_SRR14143372/README.md
@@ -59,4 +59,4 @@ intronSeeker buildReport -F --config_file  SRR14143372.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Tripidium_r_GJDA01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Tripidium_r_GJDA01.html
