+++ data/REAL_DATA/Vriesea_c_SRR500874/README.md
@@ -56,5 +56,5 @@ intronSeeker buildReport -F --config_file  SRR500874.cfg;
 
 ```
 
-HTML report is available in public directory and here https://bios4biol.pages.mia.inra.fr/intronseeker/report_Vriesea_c_GHCB01.html
+HTML report is available in public directory and here http://htmlpreview.github.io/?https://github.com/Bios4Biol/intronSeeker/blob/master/public/report_Vriesea_c_GHCB01.html
 
