### 2024-04-04 b378eeb
Update README.md
Remove GitLab CI/CD pipeline, add link to CONTRIBUTING.md file.
