### 2024-04-04 #6272 @Bios4Biol
> @SarahMaman thanks for moving the repository. Below is another point that requires your attention:
>
>     * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

Done ! Thanks for your help.
