### 2024-04-09 Bios4Biol/intronSeeker#1 @CFGrote
Installation fails here, too.

Running `setup.sh` aborted. I then ran the various perl install steps in the script manually. Here's where it failed;

```
▶ ls
Getopt-Euclid-0.4.5  Getopt-Euclid-0.4.5.tar.gz  Math-Random-MT-1.17  Math-Random-MT-1.17.tar.gz  Test-Number-Delta-1.06  Test-Number-Delta-1.06.tar.gz
(ISeeker_environment)
envs/ISeeker_environment/tmp
▶ cd Math-Random-MT-1.17
(ISeeker_environment)
ISeeker_environment/tmp/Math-Random-MT-1.17
▶ perl Makefile.PL
Generating a Unix-style Makefile
Writing Makefile for Math::Random::MT
Writing MYMETA.yml and MYMETA.json
(ISeeker_environment)
ISeeker_environment/tmp/Math-Random-MT-1.17
▶ make
Skip blib/lib/Math/Random/MT.pm (unchanged)
Running Mkbootstrap for MT ()
chmod 644 "MT.bs"
"/scratch/grotec/mambaforge/envs/ISeeker_environment/bin/perl" -MExtUtils::Command::MM -e 'cp_nonempty' -- MT.bs blib/arch/auto/Math/Random/MT/MT.bs 644
/scratch/grotec/mambaforge/envs/ISeeker_environment/bin/..//bin/x86_64-conda-linux-gnu-gcc -c   -D_REENTRANT -D_GNU_SOURCE --sysroot=/scratch/grotec/mambaforge/envs/ISeeker_environment/bin/..//x86_64-conda-linux-gnu/sysroot -fwrapv -fno-strict-aliasing -pipe -fstack-protector-strong -D_LARGEFILE_SOURCE -D_FILE_OFFSET_BITS=64 -D_FORTIFY_SOURCE=2 -O2   -DVERSION=\"1.17\" -DXS_VERSION=\"1.17\" -fPIC --sysroot=/scratch/grotec/mambaforge/envs/ISeeker_environment/bin/..//x86_64-conda-linux-gnu/sysroot "-I/scratch/grotec/mambaforge/envs/ISeeker_environment/lib/5.26.2/x86_64-linux-thread-multi/CORE"   MT.c
In file included from /scratch/grotec/mambaforge/envs/ISeeker_environment/lib/5.26.2/x86_64-linux-thread-multi/CORE/op.h:654,
                 from /scratch/grotec/mambaforge/envs/ISeeker_environment/lib/5.26.2/x86_64-linux-thread-multi/CORE/perl.h:3921,
                 from MT.xs:2:
/scratch/grotec/mambaforge/envs/ISeeker_environment/lib/5.26.2/x86_64-linux-thread-multi/CORE/reentr.h:104:17: fatal error: crypt.h: No such file or directory
  104 | #       include <crypt.h>
      |                 ^~~~~~~~~
compilation terminated.
make: *** [Makefile:336: MT.o] Error 1
(ISeeker_environment)
ISeeker_environment/tmp/Math-Random-MT-1.17                                                                                                                                                                                                                                                                      ⍉
▶
```
