### 2024-04-12 c663277
Update README.md
 JOSS REVIEW: Documentation #3 : "Open Reading Frame or open reading frame (only one should be used, maybe define and use ORF)"
