### 2024-04-12 09c4c90
Update README.md
 JOSS REVIEW: Documentation #3 
5.) Grinder file
the examples are hardly readable, please insert proper linebreaks
