### 2024-04-12 6289e75
Update README.md
 JOSS REVIEW: Documentation #3 
3.) The readme mentions a "How to use" file in doc/ directory, but there is no such file.
