### 2024-04-12 162328b
Update README.md
 JOSS REVIEW: Documentation #3 
1.) The readability could benefit from checking and fixing some language and typo issues such as:

    In order to get the correct protein sequence introns have to be removed, Retained...

-> In order to get the correct protein sequence, introns have to be removed. Retained ...
