### 2024-04-12 26e1e45
Add files via upload
 JOSS REVIEW: Documentation #3 
  The figure following "...corresponding assembled contigs." contains "ARNm", should read "mRNA".
