### 2024-04-12 Bios4Biol/intronSeeker#3 @Bios4Biol
Thank you for your comments and suggestions for improvement.

1.) They are retained by "accident" meaning due to non mature transcripts present in the cell. These transcripts have not undergone complete splicing and therefore still harbor introns. The upstream step is not data processing but a biological process called splicing.

2.) I'm working on point 2.

4.) The aligners we tested were the only two capable of handling alignments on contigs with and without the intron. We wrapped them to simplify their launch for users.

6.)  intronSeeker is a collection of CLI tools and wrappers.

7.) Indeed, automated tests have been removed.
