### 2024-04-18 Bios4Biol/intronSeeker#1 @pbordron
We thank you for the comments about creating a conda channel.

## Conda Channel

We will create a conda channel providing dependencies missing from `conda-forge` and `bioconda` channels. This will allow an easier installation of intronSeeker dependencies for the review, waiting that those packages are accepted into  `conda-forge` and `bioconda` channels.

We already have the [build recipes available](https://github.com/Bios4Biol/intronseeker-recipes)

We will naturally update the installation instructions to reflect those updates.

## `crypt.h` issue

This issue is related to some [conda-forge `sysroot` package modification](https://github.com/conda-forge/linux-sysroot-feedstock/issues/52). Our conda packages manage this build issue.

---

@CFGrote @tkchafin  We'll notify you as soon as it's available.
