### 2024-04-22 6d924f9
Comment pipeline graph
 JOSS REVIEW: Documentation #3
