### 2024-04-22 f28616d
Explain retained introns
 JOSS REVIEW: Documentation #3
