### 2024-04-22 Bios4Biol/intronSeeker#3 @Bios4Biol
Thank you @CFGrote for your comments. I've been able to integrate them into the README.
