### 2024-04-25 4b12047
Change output file name for mapping step.
In order to avoid this message:
"Error: output file(s) already exists."
