### 2024-05-02 Bios4Biol/intronSeeker#3 @CFGrote
Thanks. The README is much better now IMHO. I'm closing this one but may open another one regarding automated tests, because I believe that's a requirement for publication in JOSS.
