### 2024-05-13 1b8eb6f
(minor) Update dataSimulation.py
Remove a comment
