### 2024-05-13 5545f39
Simplify installation by relying to conda
Perl scripts in /bin can be removed
