### 2024-05-15 5bd89d2
Update README.md
Re-test with new install procedure (canal conda intronseeker)
