### 2024-05-23 Bios4Biol/intronSeeker#1 @Bios4Biol
@CFGrote @tkchafin , the installation procedure has been modified with the use of an intronSeeker conda channel. The INSTALL.md file has been updated and it is now possible to test the intronSeeker installation as well as intronSeeker. Only the last “FindEvidence” step is still being tested on our side. Thank you again.
