### 2024-07-05 #6272 @SarahMaman
Hi @mjsull and @kellyrowland,

Thank you for your message. I would just like to inform you that I will be back from vacation at the end of July. So I won't be able to read your feedback quickly. Thank you for your understanding.

Best,

Sarah

________________________________
De : mjsull ***@***.***>
Envoyé : vendredi 5 juillet 2024 09:06:54
À : openjournals/joss-reviews
Cc : Sarah Maman Haddad; Mention
Objet : Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

Hi @kellyrowland<https://github.com/kellyrowland>,

I have run into a few issues running the software, going to troubleshoot on Monday and will finalise my review then.

Best,

Mitch
