### 2024-08-27 #6272 @SarahMaman
Hello, I'm available too.

________________________________
De : Kelly L. Rowland ***@***.***>
Envoyé : lundi 26 août 2024 18:31:50
À : openjournals/joss-reviews
Cc : Sarah Maman Haddad; Mention
Objet : Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

hi @SarahMaman<https://github.com/SarahMaman> 👋 just checking back in here.
