### 2024-08-27 #6272 @tkchafin
Yes, I'll check if I can get it installed as soon as a get a chance. Swamped right now so will probably be a few weeks before I get to it.
-------- Original message --------From: "Kelly L. Rowland" ***@***.***> Date: 27/08/2024  5:17 pm  (GMT+00:00) To: openjournals/joss-reviews ***@***.***> Cc: Tyler Chafin ***@***.***>, Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing
  introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)
@mjsull @CFGrote @tkchafin are you able to pick back up your reviews here?
