### 2024-08-28 #6272 @SarahMaman
Hello @CFGrote and thank you for working on the intronSeeker installation. Following a reviewer's comment, we have slightly modified the install procedure. We will update the README as soon as possible, this week. I'll let you know when the update is complete. Yours sincerely, Sarah.

________________________________
De : Carsten Fortmann-Grote ***@***.***>
Envoyé : mardi 27 août 2024 20:35:36
À : openjournals/joss-reviews
Cc : Sarah Maman Haddad; Mention
Objet : Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

yes, working on it.
