### 2024-08-28 9b2f1fc
doc: Update installation process
