### 2024-08-29 636eb05
doc: fix links in INSTALL
