### 2024-08-29 #6272 @SarahMaman
Hello,<https://github.com/CFGrote>

The README to install intronSeeker with a Genotoul account is up to date.
Thanks,
Best regards.
Sarah Maman

________________________________
De : Sarah Maman Haddad
Envoyé : mercredi 28 août 2024 15:51:56
À : openjournals/joss-reviews; openjournals/joss-reviews
Cc : Mention
Objet : RE: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

PS: I'd just like to point out that my previous e-mail on the installation procedure only concerns installation on genotoul/genobioinfo (https://github.com/Bios4Biol/intronSeeker/blob/master/INSTALL_Genotoul.md) for our cluster's users and not local installation (https://github.com/Bios4Biol/intronSeeker/blob/master/INSTALL.md).

You can therefore test intronSeeker locally.

________________________________
De : Sarah Maman Haddad
Envoyé : mercredi 28 août 2024 14:24:35
À : openjournals/joss-reviews; openjournals/joss-reviews
Cc : Mention
Objet : RE: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

Hello @CFGrote and thank you for working on the intronSeeker installation. Following a reviewer's comment, we have slightly modified the install procedure. We will update the README as soon as possible, this week. I'll let you know when the update is complete. Yours sincerely, Sarah.

________________________________
De : Carsten Fortmann-Grote ***@***.***>
Envoyé : mardi 27 août 2024 20:35:36
À : openjournals/joss-reviews
Cc : Sarah Maman Haddad; Mention
Objet : Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

yes, working on it.
