### 2024-09-06 Bios4Biol/intronSeeker#7 @CFGrote
[JOSS Review] Running the example
JOSS issue: https://github.com/openjournals/joss-reviews/issues/6272
