### 2024-09-06 Bios4Biol/intronSeeker#7 @CFGrote
I'm running the example documented in README.md.

In step 2, the given command is relevant for the case that hisat2 was used in step 1 to produce the alignment. I used Star, so
have to change the filenames. Would be nice to have that line ready for copy'n'paste, too
