### 2024-09-06 Bios4Biol/intronSeeker#7 @CFGrote
Other than that, the example works and produces the expected results. Well done! Also big thank you for the nice results summary in the final html!
