### 2024-09-06 Bios4Biol/intronSeeker#8 @CFGrote
[JOSS Review] Paper review
JOSS issue: https://github.com/openjournals/joss-reviews/issues/6272
