### 2024-09-06 Bios4Biol/intronSeeker#8 @CFGrote
Collecting comments on the paper here:

- [x] 11: Check spelling: "aligned ~on~ to the contigs"
- [x] 21: "reads are usually assemble**d** ..."
- [x] 36: artifact (AE) vs. artefact (BE), choose and stick to one if possible
- [x] 40: ... are more difficult to **be** annotated ...
- [x] 50: there-fore -> therefore
- [x] 51: com-prise -> comprise
- [x] 57: there should be a whitespace between "from" and the URL. can you render the URL as a hyperlink?
- [x] Table 1: There's a problem with the SRA URL for GAPK01, see screenshot below
- [x] line 72: the given URL is just the generic github url, not the intronseeker repo

other than that, a very nice, informative and accurate  read!
