### 2024-09-06 #6272 @CFGrote
I have checked all boxes in my Checklist. Just a few minor comments listed in https://github.com/Bios4Biol/intronSeeker/issues/8, mostly typos and suggestions for enhanced readability.

Excellent work, I'll use it in the future!
