### 2024-09-06 3b36de7
corrections according to Joss Review
https://github.com/Bios4Biol/intronSeeker/issues/8
