### 2024-09-06 39a017e
Change "artifact"
https://github.com/Bios4Biol/intronSeeker/issues/8
