### 2024-09-06 Bios4Biol/intronSeeker#8 @Bios4Biol
Thank you @CFGrote  for your feedback. Your comments have been taken into account.
