### 2024-09-09 6227082
Update paper.md
Correct 'artefact'
