### 2024-09-09 a0fa54e
Update README.md
 [JOSS Review] Running the example #7 
Add step 2 command line for STAR
