### 2024-09-11 #6272 @mjsull
Hi Kelly, sorry ran into an issue running the program. Is it better to raise the issue here or on the github page?

Specifically, the file data/Reduced_real_dataset/Test_Cele_FindEvidence/fe_tmp/trimmed_reference.getorf has a FASTA header without an equal sign which causes an issue with

"intronSeeker/scripts/intronSearch.py", line 461, in df_lg_trim_ORF
    id    = vals[-1].split('=')[1]
