### 2024-09-24 #6272 @SarahMaman
Hello,

Thanks for this information.

Sarah
________________________________
De : Kelly L. Rowland ***@***.***>
Envoyé : lundi 23 septembre 2024 19:44:07
À : openjournals/joss-reviews
Cc : Sarah Maman Haddad; Mention
Objet : Re: [openjournals/joss-reviews] [REVIEW]: Finding and removing introns from RNA-Seq de novo assemblies with IntronSeeker (Issue #6272)

looks like things are moving along, thanks all.

just FYI @mjsull<https://github.com/mjsull> @SarahMaman<https://github.com/SarahMaman> I will be out of office starting September 30 and returning October 21. Happy to answer any questions/comments this week and after then.
