### 2024-09-26 Bios4Biol/intronSeeker#10 @mjsull
[JOSS review]
Hi authors, great paper just a couple of nitpicks in the paper that might make it a little clearer.

1.
To use it you have to provide a set of contigs resulting of a de novo transcriptome assembly and a set of RNA-Seq reads. The reads will be aligned on the contigs, splices sites will be searched and tested to check if they correspond to valid intron retention events.

The tool works by aligning user-provided reads to contigs from a de novo transcriptome assembly. Splice sites are then searched and tested to check if they correspond to valid intron retention events.

2.
[@Braunschweig2014] have shown that, even in mammals, their is widespread intron retention linked with gene expression regulation.

Their -> there

3.
Intron retention can be seen as an artefact or a biologically relevant mechanism and in both cases it is interesting to monitor.

Intron retention can be seen as an artefact or a biologically relevant mechanism; in both cases it is interesting to monitor.

4.
Contigs with introns are more difficult to be annotated because introns are splitting coding sequences (CDS) and possibly generating several shorter protein alignments rather than a unique long match.

Contigs with introns are more difficult to annotated because introns split coding sequences (CDS) and can generate several shorter protein alignments rather than a unique long match.

5.
Hereafter we present IntronSeeker a software package enabling to search and remove introns from de novo assembled RNA-Seq contigs

We present IntronSeeker; a software package that enables the user to search and remove introns from de novo assembled RNA-Seq contigs

6.
These introns generate transcript frameshifts and thus render annotation more difficult.

These introns can generate transcript frameshifts and thus render annotation more difficult.

7.
Conda has to be installed beforehand.

Conda must be installed beforehand.
