### 2024-09-26 #6272 @mjsull
Hi all,

Finished the checklist, just a couple of suggestions for the paper.

https://github.com/Bios4Biol/intronSeeker/issues/10

The tool looks very useful.

Best,

Mitch
