### 2024-09-27 1d5606b
Update paper.md
 [JOSS review] #10
