+++ paper/paper.md
@@ -35,8 +35,8 @@ bibliography: paper.bib
 # Summary 
 
 intronSeeker identifies potentially retained introns in *de novo* RNA-seq assembly in order to quantify and remove them.
-To use it you have to provide a set of contigs resulting of a *de novo* transcriptome assembly and a set of RNA-Seq reads. The reads will be aligned on the contigs, splices sites will be searched and tested to check if they correspond to valid intron retention events.
-It includes two types of RNA-seq data simulation strategies to validate the detection process and measure the false positive detection rate.
+To use it you have to provide a set of contigs resulting of a de novo transcriptome assembly and a set of RNA-Seq reads. The reads will be aligned on the contigs, splices sites will be searched and tested to check if they correspond to valid intron retention events.
+The tool works by aligning user-provided reads to contigs from a de novo transcriptome assembly. Splice sites are then searched and tested to check if they correspond to valid intron retention events.
 intronSeeker provides a list of potential intron candidates, it filters them and outputs a clean reference without introns in Fasta format.
 
 # Introduction
@@ -45,21 +45,21 @@ Short read RNA sequencing (RNA-Seq) is now routinely used to study gene expressi
 
 In order not to sequence very abundant ribosomal RNAs, commonly used protocols include oligo(dT) transcript enrichment. Transcript poly-adhenylation and splicing take place in the nucleus before transfer in the cytoplasm. PolyA tail selection retrieves mainly mature spliced transcripts. Introns spanning reads are still found in most RNA-Seq sets as shown by  [@Ameur2011].
 
-Intron retention is also a known biological gene regulation or alternative splicing mechanism. In plants it has been shown to increase the number of transcript splicing forms  [@Jia2020].  [@Braunschweig2014] have shown that, even in mammals, their is widespread intron retention linked with gene expression regulation.
+Intron retention is also a known biological gene regulation or alternative splicing mechanism. In plants it has been shown to increase the number of transcript splicing forms  [@Jia2020].  [@Braunschweig2014] have shown that, even in mammals, there is widespread intron retention linked with gene expression regulation.
 
 Reads located on intron/exon boundaries harbor specific splice motifs. These splice motifs are di-nucleotides located at both intron ends. They can be canonical (GT/AG) or not. Canonical motifs are found in most site of most species.
 
-Intron retention can be seen as an artefact or a biologically relevant mechanism and in both cases it is interesting to monitor. This is easy with a reference genome assembly because one can quantify reads aligned in introns. This is more complicated in the de novo approach in which the assembler can produce contigs with and without intron for the same transcript.
+Intron retention can be seen as an artefact or a biologically relevant mechanism; in both cases it is interesting to monitor. This is easy with a reference genome assembly because one can quantify reads aligned in introns. This is more complicated in the de novo approach in which the assembler can produce contigs with and without intron for the same transcript.
 
-Contigs with introns are more difficult to be annotated because introns are splitting coding sequences (CDS) and possibly generating several shorter protein alignments rather than a unique long match.
+Contigs with introns are more difficult to annotated because introns split coding sequences (CDS) and can generate several shorter protein alignments rather than a unique long match.
 
 Canonical splice motifs presence, number of reads splicing at the same location, the minimum number of splice events and overlaps can be used to reduce faulty detection.
 
-Hereafter we present IntronSeeker a software package enabling to search and remove introns from de novo assembled RNA-Seq contigs.
+We present IntronSeeker; a software package that enables the user to search and remove introns from de novo assembled RNA-Seq contigs.
 
 # Statement of need
 
[…]
