### 2024-09-27 Bios4Biol/intronSeeker#10 @Bios4Biol
Hi @mjsull , We would like to thank you for your suggestions for improvement and for the time you have spent on intronSeeker.
