### 2024-10-25 #6272 @Bios4Biol
Hello @kellyrowland
Thanks for your message, please find the requested information :
- intronSeeker version : v1.0.0
- archive DOI - https://zenodo.org/records/13991675 : 10.5281/zenodo.13991675
