### 2024-10-30 #6272 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
