### 2024-10-31 #6272 @Kevin-Mattheus-Moerman
@SarahMaman  as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process some final checks:

* One reviewer @tkchafin has not but two others have.
- [x] Software license tag listed here matches a tagged release

- [ ] Please ensure the archive title matches the paper title
- [ ] Please ensure the archive listed license version matches your license tag i.e. set it to be `v1.0.0` (making sure to include the `v` as well).
