### 2024-11-01 #6272 @Kevin-Mattheus-Moerman
@SarahMaman @Bios4Biol thanks. Several issues remain:

- [ ] Make sure the authors listed on the archive match the paper authors
- [ ] Make sure the license tag listed on the archive matches the software release tag used here, i.e. `v1.0.0`.

Note that to make these changes you do not need to create a new archive. Instead you can manually edit the current one.
