### 2024-11-14 #6272 @Bios4Biol
Hello @Kevin-Mattheus-Moerman
The article does not yet seem to be available on JOSS and the DOI (https://doi.org/10.21105/joss.06272) has not yet been published. Should we take action and just be patient?
Regards
