### 2024-11-15 #6272 @Kevin-Mattheus-Moerman
@Bios4Biol Apologies for the delaying in checking in here. I see you've updated author set listed on the archive, thanks. However we are now waiting for you to edit the version listed on the archive. This should match the version tag you created for this review (https://github.com/Bios4Biol/intronSeeker/tags) and which is listed here, i.e. `v1.0.0`. Currently the archive says `v2`.
There is no need to create a new archive, you can log in over at ZENODO and manually edit the archive. Let me know when this is done.
